third_party/WebKit/LayoutTests/typedcssom/stylevalue-normalization/normalize-ident.html

<!doctype html>
<meta charset="utf-8">
<title>Identifier normalization tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#normalize-ident">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/testhelper.js"></script>
<body>
<script>
'use strict';

test(() => {
  assert_matches(CSSStyleValue.parse('width', 'auto'),
      Matchers.KeywordValue('auto'));
}, 'CSS identifiers are normalized from String to CSSKeywordValues');

test(t => {
  assert_matches(
      createInlineStyleMap(t, 'width: auto').get('width'),
      Matchers.KeywordValue('auto'));
}, 'CSS identifiers are normalized from CSSOM to CSSKeywordValues');

</script>
